domain/pyproject.toml
[tool.poetry]
name = "app-domain"
version = "0.1.0"
description = ""
authors = ["gersonUb <gersoncabj@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app_domain", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.12.5"
email-validator = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
ruff = "^0.14.14"
pytest-cov = "^7.0.0"
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"